Error: The global `ok` is removed in QUnit 2.0.
ember-cli has updated to QUnit 2.0 in it's 2.9.0 release.

QUnit 2.0 has dropped global assertions.
Therefore tests are throwing if ember-cli-qunit 3.0 is used, which is
included in ember-cli since 2.9.0 release.

To drop global assertions we have to pass assert to test helper as suggested
by examples in ember guide:
https://guides.emberjs.com/v2.10.0/testing/acceptance/#toc_custom-test-helpers

This is a breaking change. E.q.
  expectComponent(componentName, count, options)
has to be changed to
  expectComponent(assert, componentName, count, options)